requirements.txt
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
seaborn==0.13.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.67.1
